renamed test classes
