+++ src/test/java/ru/practikum/tests/ConstructorTests.java
@@ -1,0 +1,4 @@
+package ru.practikum.tests;
+
+public class ConstructorTests {
+}
