+++ src/test/java/ru/practikum/tests/SuccessfulRegistrationTests.java
@@ -1,16 +1,17 @@
 package ru.practikum.tests;
 
 import io.qameta.allure.Description;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
+import ru.practikum.pages.LoginPage;
 import ru.practikum.pages.RegisterPage;
+import org.apache.commons.lang3.RandomStringUtils;
 
 @RunWith(Parameterized.class)
-public class IncorrectPasswordErrorTest {
+public class SuccessfulRegistrationTests {
     @Rule
     public DriverFactory driverFactory = new DriverFactory();
 
@@ -19,7 +20,7 @@ public class IncorrectPasswordErrorTest {
     private final String password;
     private final int passwordSymbolsNumber;
 
-    public IncorrectPasswordErrorTest(String userName, String email, String password, int passwordSymbolsNumber) {
+    public SuccessfulRegistrationTests(String userName, String email, String password, int passwordSymbolsNumber) {
         this.userName = userName;
         this.email = email;
         this.password = password;
@@ -29,22 +30,21 @@ public class IncorrectPasswordErrorTest {
     @Parameterized.Parameters(name = " Symbols number in password - {3}")
     public static Object[][] getOrderData() {
         return new Object[][] {
-                {"Иван", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", "", 0},
-                {"Мария", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(1), 1},
-                {"Анна", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(3), 3},
-                {"Пётр", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(4), 4},
-                {"Вера", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(5), 5},
+                {"Иван", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(6), 6},
+                {"Мария", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(7), 7},
+                {"Анна", RandomStringUtils.randomAlphanumeric(8) + "@test.ru", RandomStringUtils.randomAlphanumeric(10), 10}
         };
     }
 
     @Test
-    @Description("Test an appearance of an error message when typing incorrect password in field")
-    public void errorMessageOnIncorrectPasswordTest() throws InterruptedException {
+    @Description("Register user with correct email, password and name")
+    public void registerUserSuccessfullyTest() throws InterruptedException {
         WebDriver driver = driverFactory.getDriver();
 
         RegisterPage registerPageObj = new RegisterPage(driver);
         registerPageObj.openRegisterPage();
         registerPageObj.registerUser(userName, email, password);
-        registerPageObj.checkErrorMessageIsDisplayed();
+        LoginPage loginPageObj = new LoginPage(driver);
+        loginPageObj.checkLoginButtonIsDisplayed();
     }
 }
